nettool/requirements.txt
requests>=2.28.0
python-nmap>=0.7.1
scapy>=2.5.0
dnspython>=2.3.0

